refactor: collectible list
