+++ packages/mask/src/extension/options-page/DashboardComponents/CollectibleList/CollectibleListContext.tsx
@@ -1,0 +1,5 @@
+import { createContext, } from 'react'
+
+export const CollectibleListContext = createContext<{
+    collectiblesRetry: () => void
+}>(null!)
