+++ packages/mask/src/extension/options-page/DashboardDialogs/Wallet/TransferNFT.tsx
@@ -4,7 +4,7 @@ import { useCallback, useContext, useMemo, useState } from 'react'
 import { EthereumAddress } from 'wallet.ts'
 import type { ChainId, SchemaType } from '@masknet/web3-shared-evm'
 import { useI18N } from '../../../../utils/index.js'
-import { CollectibleContext } from '../../DashboardComponents/CollectibleList/index.js'
+import { CollectibleListContext } from '../../DashboardComponents/CollectibleList/index.js'
 import { DashboardDialogCore, DashboardDialogWrapper, WrappedDialogProps } from '../Base.js'
 import type { NonFungibleToken } from '@masknet/web3-shared-base'
 import { useTokenTransferCallback } from '@masknet/plugin-infra/web3-evm'
@@ -36,7 +36,7 @@ export function DashboardWalletTransferDialogNFT(
     const { classes } = useTransferDialogStylesNFT()
 
     const [address, setAddress] = useState('')
-    const { collectiblesRetry } = useContext(CollectibleContext)
+    const { collectiblesRetry } = useContext(CollectibleListContext)
 
     // #region transfer tokens
     const [{ loading }, transferCallback] = useTokenTransferCallback(token.schema, token.address)
